fix: enforce bootstrap dispatch ownership
