+++ backend/app/services/market_rs_activation_coverage.py
@@ -5,13 +5,12 @@ from __future__ import annotations
 from dataclasses import dataclass
 from datetime import date, timedelta
 
-from app.services.market_calendar_service import MarketCalendarService
 from app.services.group_rank_history_policy import (
     DEFAULT_CALENDAR_DAY_GROUP_RANK_HISTORY_LOOKBACK_DAYS,
 )
+from app.services.market_calendar_service import MarketCalendarService
 from app.services.market_rs_rollout_contracts import normalize_rollout_market
 from app.services.rrg_service import MIN_TAIL_WEEKS
-
 
 RRG_MINIMUM_LOOKBACK_DAYS = MIN_TAIL_WEEKS * 7
 MARKET_RS_ACTIVATION_LOOKBACK_DAYS = max(
@@ -32,6 +31,18 @@ class MarketRsActivationCoverage:
     through_date: date
     required_dates: tuple[date, ...]
 
+    def __post_init__(self) -> None:
+        normalized = normalize_rollout_market(self.market)
+        required_dates = tuple(self.required_dates)
+        if not required_dates or required_dates[-1] != self.through_date:
+            raise ValueError(
+                "Guarded activation date must be the final required trading date."
+            )
+        if tuple(sorted(set(required_dates))) != required_dates:
+            raise ValueError("Guarded activation dates must be unique and increasing.")
+        object.__setattr__(self, "market", normalized)
+        object.__setattr__(self, "required_dates", required_dates)
+
     @property
     def start_date(self) -> date:
         return self.required_dates[0]
@@ -43,7 +54,7 @@ class MarketRsActivationCoverage:
         calendar_service: MarketCalendarService,
         market: str,
         through_date: date,
-    ) -> "MarketRsActivationCoverage":
+    ) -> MarketRsActivationCoverage:
         normalized = normalize_rollout_market(market)
         required_dates = tuple(
             calendar_service.trading_days(
