+++ backend/app/wiring/bootstrap.py
@@ -2,36 +2,46 @@
 
 from __future__ import annotations
 
-from contextvars import ContextVar, Token
+from collections.abc import Callable, Iterator
+from contextvars import Token
 from dataclasses import dataclass
 from threading import RLock
-from typing import TYPE_CHECKING, Callable, Iterator, TypeAlias
+from typing import TYPE_CHECKING, TypeAlias
 
 from fastapi import Request
 from sqlalchemy.orm import Session
 
 from app.database import SessionLocal
 from app.domain.scanning.ports import StockDataProvider, TaskDispatcher
-from app.services.job_backend import JobBackend, CeleryJobBackend
+from app.services.job_backend import CeleryJobBackend, JobBackend
 from app.services.redis_pool import get_redis_client
-from app.wiring.use_case_factories import (
-    get_build_daily_snapshot_use_case as get_build_daily_snapshot_use_case,
-    get_compare_feature_runs_use_case as get_compare_feature_runs_use_case,
-    get_create_scan_use_case as get_create_scan_use_case,
-    get_create_scan_use_case_without_freshness_gate as get_create_scan_use_case_without_freshness_gate,
-    get_explain_stock_use_case as get_explain_stock_use_case,
-    get_export_scan_results_use_case as get_export_scan_results_use_case,
-    get_get_filter_options_use_case as get_get_filter_options_use_case,
-    get_get_peers_use_case as get_get_peers_use_case,
-    get_get_scan_results_use_case as get_get_scan_results_use_case,
-    get_get_scan_symbols_use_case as get_get_scan_symbols_use_case,
-    get_get_setup_details_use_case as get_get_setup_details_use_case,
-    get_get_single_result_use_case as get_get_single_result_use_case,
-    get_list_feature_runs_use_case as get_list_feature_runs_use_case,
-    get_run_bulk_scan_use_case as get_run_bulk_scan_use_case,
+from app.wiring import runtime_context
+from app.wiring import use_case_factories as _use_case_factories
+
+get_build_daily_snapshot_use_case = (
+    _use_case_factories.get_build_daily_snapshot_use_case
 )
+get_compare_feature_runs_use_case = (
+    _use_case_factories.get_compare_feature_runs_use_case
+)
+get_create_scan_use_case = _use_case_factories.get_create_scan_use_case
+get_create_scan_use_case_without_freshness_gate = (
+    _use_case_factories.get_create_scan_use_case_without_freshness_gate
+)
+get_explain_stock_use_case = _use_case_factories.get_explain_stock_use_case
+get_export_scan_results_use_case = _use_case_factories.get_export_scan_results_use_case
+get_get_filter_options_use_case = _use_case_factories.get_get_filter_options_use_case
+get_get_peers_use_case = _use_case_factories.get_get_peers_use_case
+get_get_scan_results_use_case = _use_case_factories.get_get_scan_results_use_case
+get_get_scan_symbols_use_case = _use_case_factories.get_get_scan_symbols_use_case
+get_get_setup_details_use_case = _use_case_factories.get_get_setup_details_use_case
+get_get_single_result_use_case = _use_case_factories.get_get_single_result_use_case
+get_list_feature_runs_use_case = _use_case_factories.get_list_feature_runs_use_case
+get_run_bulk_scan_use_case = _use_case_factories.get_run_bulk_scan_use_case
 
 if TYPE_CHECKING:
+    from app.domain.scanning.ports import MarketRsReader
+    from app.infra.db.repositories.market_rs_repo import MarketRsRunRepository
     from app.infra.db.uow import SqlUnitOfWork
     from app.infra.providers.stock_data import DataPrepStockDataProvider
     from app.scanners.scan_orchestrator import ScanOrchestrator
@@ -40,30 +50,26 @@ if TYPE_CHECKING:
     from app.services.canonical_group_ranking_service import (
         CanonicalGroupRankingService,
     )
+    from app.services.daily_price_bundle_service import DailyPriceBundleService
     from app.services.data_source_service import DataSourceService
     from app.services.eps_rating_service import EPSRatingService
     from app.services.finviz_service import FinvizService
     from app.services.fundamentals_cache_service import FundamentalsCacheService
     from app.services.github_release_sync_service import GitHubReleaseSyncService
-    from app.services.hybrid_fundamentals_service import HybridFundamentalsService
-    from app.services.ibd_group_rank_service import IBDGroupRankService
-    from app.services.llm.groq_key_manager import GroqKeyManager
-    from app.services.llm.zai_key_manager import ZAIKeyManager
-    from app.services.daily_price_bundle_service import DailyPriceBundleService
-    from app.services.market_calendar_service import MarketCalendarService
-    from app.services.market_rs_inputs import MarketRsInputLoader
-    from app.domain.scanning.ports import MarketRsReader
-    from app.services.market_rs_snapshot_service import MarketRsSnapshotService
-    from app.services.market_rs_rollout_executor import MarketRsActivationExecutor
-    from app.services.market_rs_rollout_service import MarketRsRolloutService
     from app.services.group_rank_snapshot_coordinator import (
         GroupRankSnapshotCoordinator,
     )
     from app.services.group_rank_snapshot_reader import GroupRankSnapshotReader
-    from app.wiring.market_rs_services import MarketRsServices
+    from app.services.hybrid_fundamentals_service import HybridFundamentalsService
+    from app.services.ibd_group_rank_service import IBDGroupRankService
+    from app.services.llm.groq_key_manager import GroqKeyManager
+    from app.services.llm.zai_key_manager import ZAIKeyManager
+    from app.services.market_calendar_service import MarketCalendarService
+    from app.services.market_rs_inputs import MarketRsInputLoader
+    from app.services.market_rs_rollout_executor import MarketRsActivationExecutor
+    from app.services.market_rs_rollout_service import MarketRsRolloutService
+    from app.services.market_rs_snapshot_service import MarketRsSnapshotService
     from app.services.point_in_time_universe_service import PointInTimeUniverseService
-    from app.infra.db.repositories.market_rs_repo import MarketRsRunRepository
-    from app.wiring.canonical_rs_runtime import CanonicalRsRuntime
     from app.services.price_cache_service import PriceCacheService
     from app.services.provider_snapshot_service import ProviderSnapshotService
     from app.services.rate_limiter import RedisRateLimiter
@@ -76,6 +82,8 @@ if TYPE_CHECKING:
     from app.services.yfinance_service import YFinanceService
     from app.tasks.data_fetch_lock import DataFetchLock
     from app.tasks.workload_coordination import WorkloadCoordination
+    from app.wiring.canonical_rs_runtime import CanonicalRsRuntime
+    from app.wiring.market_rs_services import MarketRsServices
 
 
 SessionFactory: TypeAlias = Callable[[], Session]
@@ -188,6 +196,12 @@ class RuntimeServices:
         if self._group_rank_service is None:
             with self._init_lock:
                 if self._group_rank_service is None:
+                    from app.scanners.criteria.relative_strength import (
+                        RelativeStrengthCalculator,
+                    )
+                    from app.services.group_rank_historical_calculator import (
+                        GroupRankHistoricalCalculator,
+                    )
                     from app.services.group_rank_input_loader import (
                         GroupRankInputLoader,
                     )
@@ -195,9 +209,6 @@ class RuntimeServices:
                         IBDIndustryTaxonomySource,
                         SqlGroupRankMarketCapSource,
                         StockUniverseGroupRankSource,
-                    )
-                    from app.services.group_rank_historical_calculator import (
-                        GroupRankHistoricalCalculator,
                     )
                     from app.services.group_rank_legacy_adapter import (
                         LegacyGroupRankPrefetchAdapter,
@@ -209,9 +220,6 @@ class RuntimeServices:
                         GroupRankingRepository,
                     )
                     from app.services.ibd_group_rank_service import IBDGroupRankService
-                    from app.scanners.criteria.relative_strength import (
-                        RelativeStrengthCalculator,
-                    )
 
                     cache_bundle = self.cache_bundle()
                     input_loader = GroupRankInputLoader(
@@ -397,10 +405,10 @@ class RuntimeServices:
         if self._github_release_sync_service is None:
             with self._init_lock:
                 if self._github_release_sync_service is None:
+                    from app.config import settings
                     from app.services.github_release_sync_service import (
                         GitHubReleaseSyncService,
                     )
-                    from app.config import settings
 
                     self._github_release_sync_service = GitHubReleaseSyncService(
                         api_base=settings.github_data_api_base,
@@ -596,14 +604,6 @@ class RuntimeServices:
             self._scan_orchestrator = None
 
 
-_runtime_services_ctx: ContextVar[RuntimeServices | None] = ContextVar(
-    "runtime_services_ctx",
-    default=None,
-)
-_process_runtime_services_lock = RLock()
-_process_runtime_services = None  # type: RuntimeServices | None
-
-
 def build_runtime_services(
     *,
     session_factory: SessionFactory = SessionLocal,
@@ -621,16 +621,12 @@ def set_runtime_services(
 
     Set ``bind_process`` only at process lifecycle boundaries.
     """
-    global _process_runtime_services
-    if bind_process:
-        with _process_runtime_services_lock:
-            _process_runtime_services = runtime
-    return _runtime_services_ctx.set(runtime)
+    return runtime_context.set_runtime_services(runtime, bind_process=bind_process)
 
 
 def reset_runtime_services(token: Token[RuntimeServices | None]) -> None:
     """Restore runtime services context to a previous token."""
-    _runtime_services_ctx.reset(token)
+    runtime_context.reset_runtime_services(token)
 
 
 def _provision_market_rs_formula_pointers(
@@ -654,51 +650,29 @@ def initialize_process_runtime_services(
     force: bool = False,
 ) -> RuntimeServices:
     """Ensure one process-scoped runtime container exists and bind it to context."""
-    global _process_runtime_services
-    with _process_runtime_services_lock:
-        if _process_runtime_services is None or force:
-            _provision_market_rs_formula_pointers(session_factory)
-            _process_runtime_services = build_runtime_services(
-                session_factory=session_factory
-            )
-        runtime = _process_runtime_services
-    set_runtime_services(runtime, bind_process=True)
-    return runtime
+
+    def _factory() -> RuntimeServices:
+        _provision_market_rs_formula_pointers(session_factory)
+        return build_runtime_services(session_factory=session_factory)
+
+    return runtime_context.initialize_process_runtime_services(
+        _factory,
+        force=force,
+    )
 
 
 def clear_runtime_services() -> None:
     """Clear runtime services from both process state and current context."""
-    global _process_runtime_services
-    with _process_runtime_services_lock:
-        _process_runtime_services = None
-    _runtime_services_ctx.set(None)
+    runtime_context.clear_runtime_services()
 
 
 def get_runtime_services(request: Request) -> RuntimeServices:
     """FastAPI dependency getter for process runtime services."""
-    runtime = getattr(request.app.state, "runtime_services", None)
-    if runtime is None:
-        raise RuntimeError(
-            "RuntimeServices are not initialized on app.state.runtime_services"
-        )
-    return runtime
+    return runtime_context.request_runtime_services(request)
 
 
 def _resolve_runtime_services(request: Request | None = None) -> RuntimeServices:
-    if request is not None:
-        request_runtime = getattr(request.app.state, "runtime_services", None)
-        if request_runtime is not None:
-            return request_runtime
-    context_runtime = _runtime_services_ctx.get()
-    if context_runtime is not None:
-        return context_runtime
-    with _process_runtime_services_lock:
-        if _process_runtime_services is not None:
-            return _process_runtime_services
-    raise RuntimeError(
-        "RuntimeServices are not initialized for this context. "
-        "Call initialize_process_runtime_services() at process startup."
-    )
+    return runtime_context.resolve_runtime_services(request)
 
 
 def get_uow() -> Iterator[SqlUnitOfWork]:
@@ -904,7 +878,7 @@ def get_scan_orchestrator() -> ScanOrchestrator:
 
 def _reset_singletons_for_tests() -> None:
     """Reset runtime-scoped singletons for test isolation."""
-    runtime = _runtime_services_ctx.get()
+    runtime = runtime_context.current_runtime_services()
     if runtime is not None:
         runtime.reset_for_tests()
     clear_runtime_services()
