+++ backend/tests/unit/test_bootstrap_run_manifest.py
@@ -6,9 +6,8 @@ import pytest
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from app.database import Base
-
 import app.models.app_settings  # noqa: F401
+from app.database import Base
 
 
 def test_bootstrap_run_manifest_repository_round_trips_market_task_ids():
@@ -183,8 +182,9 @@ def test_dispatch_update_preserves_consumed_pending_markets() -> None:
         engine.dispose()
 
 
-def test_new_dispatch_can_record_a_new_fresh_classification() -> None:
+def test_repository_rejects_a_new_dispatch_while_the_current_one_is_active() -> None:
     from app.services.bootstrap_run_manifest import (
+        BootstrapAlreadyRunning,
         BootstrapRunManifest,
         BootstrapRunManifestRepository,
     )
@@ -204,27 +204,70 @@ def test_new_dispatch_can_record_a_new_fresh_classification() -> None:
             ),
         )
 
-        repository.begin_dispatch(
-            db,
-            BootstrapRunManifest.create(
-                primary_market="US",
-                enabled_markets=("US",),
-                fresh_install=True,
-                dispatch_id="dispatch-b",
-            ),
-        )
+        with pytest.raises(BootstrapAlreadyRunning, match="dispatch-a"):
+            repository.begin_dispatch(
+                db,
+                BootstrapRunManifest.create(
+                    primary_market="US",
+                    enabled_markets=("US",),
+                    fresh_install=True,
+                    dispatch_id="dispatch-b",
+                ),
+            )
 
-        assert repository.load(db).fresh_install is True
+        assert repository.load(db).dispatch_id == "dispatch-a"
     finally:
         db.close()
         engine.dispose()
 
 
-def test_repository_rejects_update_from_an_older_dispatch_generation() -> None:
+def test_repository_can_supersede_a_legacy_manifest_without_generation_ownership() -> (
+    None
+):
+    from app.models.app_settings import AppSetting
+    from app.services.bootstrap_run_manifest import (
+        BOOTSTRAP_RUN_KEY,
+        BootstrapRunManifest,
+        BootstrapRunManifestRepository,
+    )
+
+    engine = create_engine("sqlite:///:memory:")
+    Base.metadata.create_all(engine)
+    db = sessionmaker(bind=engine)()
+    try:
+        db.add(
+            AppSetting(
+                key=BOOTSTRAP_RUN_KEY,
+                value=(
+                    '{"primary_market":"US","enabled_markets":["US"],'
+                    '"dispatch_id":"legacy-dispatch","queue_state":"queued"}'
+                ),
+            )
+        )
+        db.commit()
+
+        BootstrapRunManifestRepository().begin_dispatch(
+            db,
+            BootstrapRunManifest.create(
+                primary_market="HK",
+                enabled_markets=("HK",),
+                dispatch_id="dispatch-current",
+                queue_state="queueing",
+            ),
+        )
+
+        assert BootstrapRunManifestRepository().load(db).dispatch_id == (
+            "dispatch-current"
+        )
+    finally:
+        db.close()
+        engine.dispose()
+
+
+def test_repository_allows_a_new_dispatch_after_the_current_one_is_terminal() -> None:
     from app.services.bootstrap_run_manifest import (
         BootstrapRunManifest,
         BootstrapRunManifestRepository,
-        StaleBootstrapDispatch,
     )
 
     engine = create_engine("sqlite:///:memory:")
@@ -238,18 +281,63 @@ def test_repository_rejects_update_from_an_older_dispatch_generation() -> None:
             dispatch_id="dispatch-a",
             queue_state="queueing",
         )
+        completed = replace(first, queue_state="completed")
         second = replace(first, dispatch_id="dispatch-b")
         repository.begin_dispatch(db, first)
+        repository.update_dispatch(
+            db,
+            dispatch_id="dispatch-a",
+            transform=lambda _manifest: completed,
+        )
         repository.begin_dispatch(db, second)
 
-        with pytest.raises(StaleBootstrapDispatch, match="dispatch-a"):
-            repository.update_dispatch(
-                db,
-                dispatch_id="dispatch-a",
-                transform=lambda manifest: replace(manifest, queue_state="queued"),
-            )
-
         assert repository.load(db).dispatch_id == "dispatch-b"
+    finally:
+        db.close()
+        engine.dispose()
+
+
+def test_queued_update_reconciles_a_market_that_completed_before_task_id_recording() -> (
+    None
+):
+    from app.services.bootstrap_run_manifest import (
+        BootstrapQueueState,
+        BootstrapRunManifest,
+        BootstrapRunManifestRepository,
+    )
+
+    engine = create_engine("sqlite:///:memory:")
+    Base.metadata.create_all(engine)
+    db = sessionmaker(bind=engine)()
+    try:
+        repository = BootstrapRunManifestRepository()
+        repository.begin_dispatch(
+            db,
+            BootstrapRunManifest.create(
+                primary_market="US",
+                enabled_markets=("US",),
+                dispatch_id="dispatch-a",
+                queue_state="queueing",
+            ),
+        )
+        repository.finish_market(
+            db,
+            dispatch_id="dispatch-a",
+            market="US",
+            succeeded=True,
+        )
+        repository.update_dispatch(
+            db,
+            dispatch_id="dispatch-a",
+            transform=lambda current: replace(
+                current,
+                primary_task_id="task-us",
+                market_task_ids={"US": "task-us"},
+                queue_state="queued",
+            ),
+        )
+
+        assert repository.load(db).queue_state == BootstrapQueueState.COMPLETED
     finally:
         db.close()
         engine.dispose()
