+++ backend/tests/unit/test_market_rs_activation_coverage.py
@@ -1,6 +1,8 @@
 from datetime import date, timedelta
 from unittest.mock import MagicMock
 
+import pytest
+
 from app.services.market_rs_activation_coverage import (
     MARKET_RS_ACTIVATION_LOOKBACK_DAYS,
     MarketRsActivationCoverage,
@@ -27,3 +29,18 @@ def test_activation_coverage_is_the_exact_bounded_trading_calendar() -> None:
         through_date - timedelta(days=MARKET_RS_ACTIVATION_LOOKBACK_DAYS),
         through_date,
     )
+
+
+@pytest.mark.parametrize(
+    "required_dates",
+    [(), (date(2026, 7, 28),)],
+)
+def test_activation_coverage_rejects_an_incomplete_required_calendar(
+    required_dates,
+) -> None:
+    with pytest.raises(ValueError, match="activation date"):
+        MarketRsActivationCoverage(
+            market="US",
+            through_date=date(2026, 7, 29),
+            required_dates=required_dates,
+        )
